requirements.txt
emoji==2.2.0
Flask==2.2.2
pytest==7.2.1
requests==2.28.2
pydantic==1.10.4
jsonschema==4.17.3
allure-pytest==2.12.0
